preprocessing and HW optimised
